refactor: Entity에서 Dto 의존하는 문제 해결

+++ product-service/src/main/java/com/sasstyle/productservice/service/ProductService.java
@@ -1,9 +1,11 @@
 package com.sasstyle.productservice.service;
 
+import com.sasstyle.productservice.client.UserServiceClient;
 import com.sasstyle.productservice.controller.dto.*;
 import com.sasstyle.productservice.entity.Category;
 import com.sasstyle.productservice.entity.Product;
-import com.sasstyle.productservice.entity.ProductDetail;
+import com.sasstyle.productservice.entity.ProductImage;
+import com.sasstyle.productservice.entity.ProductProfile;
 import com.sasstyle.productservice.repository.CategoryRepository;
 import com.sasstyle.productservice.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
@@ -11,9 +13,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.stream.Collectors;
 
 import static java.util.Objects.isNull;
 
@@ -24,6 +29,7 @@ public class ProductService {
 
     private final CategoryRepository categoryRepository;
     private final ProductRepository productRepository;
+    private final UserServiceClient userServiceClient;
 
     public Product findById(Long productId) {
         return productRepository.findById(productId)
@@ -44,12 +50,12 @@ public class ProductService {
         return productRepository.findProducts(pageable);
     }
 
-    public Page<ProductResponse> searchInQuery(List<Long> categoryIds, Pageable pageable) {
-        return productRepository.searchInQuery(categoryIds, pageable);
-    }
-
     public Page<ProductResponse> search(ProductSearch productSearch, Pageable pageable) {
         return productRepository.search(productSearch, pageable);
+    }
+
+    public Page<ProductResponse> searchInQuery(List<Long> categoryIds, Pageable pageable) {
+        return productRepository.searchInQuery(categoryIds, pageable);
     }
 
     public List<ProductAutoCompleteResponse> autocomplete(ProductSearch productSearch, Pageable pageable) {
@@ -57,33 +63,85 @@ public class ProductService {
     }
 
     @Transactional
-    public Long createProduct(ProductRequest request) {
+    public Long createProduct(String userId, ProductRequest request) {
         // 카테고리 조회
-        Category category = categoryRepository.findById(request.getCategoryId())
+        Category findCategory = categoryRepository.findById(request.getCategoryId())
                 .orElseThrow(() -> new NoSuchElementException("카테고리를 찾을 수 없습니다."));
 
-        // 상품 생성
-        Product product = Product.create(category, request);
+        // 사용자 조회
+        UserResponse userResponse = userServiceClient.findByUserId(userId);
 
-        // 상품 디테일 이미지 생성
-        request.getDetailImages().stream()
-                .map(ProductDetail::new)
-                .forEach(product::addDetailImage);
+        // 상품 프로필 생성
+        ProductProfile productProfile = ProductProfile.builder()
+                .profileUrl(request.getProfileUrl())
+                .build();
+
+        // 상품 상세 이미지 생성
+        List<ProductImage> productImages = new ArrayList<>();
+        if (isUploadImages(request.getImages())) {
+            productImages = request.getImages().stream()
+                    .map(ProductImage::new)
+                    .collect(Collectors.toList());
+        }
+
+        // 상품 생성
+        Product product = Product.builder()
+                .category(findCategory)
+                .productProfile(productProfile)
+                .userId(userId)
+                .brandName(userResponse.getName())
+                .name(request.getName())
+                .price(request.getPrice())
+                .stockQuantity(request.getStockQuantity())
+                .topDescription(request.getTopDescription())
+                .bottomDescription(request.getBottomDescription())
+                .productImages(productImages)
+                .build();
 
         return productRepository.save(product).getId();
     }
 
     @Transactional
-    public void updateProduct(Long productId, ProductUpdateRequest request) {
-        Product product = findById(productId);
+    public void updateProduct(String userId, Long productId, ProductUpdateRequest request) {
+        Product product = findProduct(productId);
 
-        product.update(request);
+        if (isNotValidSeller(userId, product.getUserId())) {
+            throw new IllegalArgumentException("상품 판매자와 로그인한 사용자가 일치하지 않습니다.");
+        }
+
+        // 상품 프로필 업데이트
+        if (hasProfileUrl(request.getProfileUrl())) {
+            ProductProfile productProfile = product.getProductProfile();
+            productProfile.setProfileUrl(request.getProfileUrl());
+        }
+
+        product.update(request.getName(),
+                request.getPrice(),
+                request.getStockQuantity(),
+                request.getTopDescription(),
+                request.getBottomDescription());
     }
 
     @Transactional
-    public void deleteProduct(Long productId) {
+    public void deleteProduct(String userId, Long productId) {
         Product product = findById(productId);
 
+        if (isNotValidSeller(userId, product.getUserId())) {
+            throw new IllegalArgumentException("상품 판매자와 로그인한 사용자가 일치하지 않습니다.");
+        }
+
         productRepository.delete(product);
+    }
+
+    private boolean isNotValidSeller(String userId, String productUserId) {
+        return !userId.equals(productUserId);
+    }
+
+    private boolean hasProfileUrl(String profileUrl) {
+        return StringUtils.hasText(profileUrl);
+    }
+
+    private boolean isUploadImages(List<String> images) {
+        return images != null && !images.isEmpty();
     }
 }
